requirements.txt
langchain
langchain-openai
langchain-ollama
langchain-core
python-dotenv
